+++ modules/terracotta-core/src/main/kotlin/io/github/beduality/terracotta/core/model/TerracottaProject.kt
@@ -18,8 +18,8 @@ data class TerracottaProject(
     val description: String,
     /** Released versions of the project. */
     val versions: List<TerracottaVersion>,
-    /** Search tags. */
-    val tags: List<String>,
+    /** Project categories. */
+    val categories: TerracottaProjectCategories,
     /** SPDX license identifier. */
     val license: String,
     /** Optional URL to the full license text. */