+++ modules/terracotta-gradle-plugin/src/main/kotlin/io/github/beduality/terracotta/gradle/TerracottaPlanTask.kt
@@ -5,6 +5,7 @@ import io.github.beduality.terracotta.core.diff.OperationPreprocessor
 import io.github.beduality.terracotta.core.model.TerracottaEnvironment
 import io.github.beduality.terracotta.core.model.TerracottaGalleryItem
 import io.github.beduality.terracotta.core.model.TerracottaProject
+import io.github.beduality.terracotta.core.model.TerracottaProjectCategories
 import io.github.beduality.terracotta.core.model.TerracottaProjectLinks
 import io.github.beduality.terracotta.core.model.releasetype.TerracottaReleaseType
 import io.github.beduality.terracotta.core.model.version.TerracottaVersion
@@ -51,8 +52,8 @@ abstract class TerracottaPlanTask : DefaultTask() {
     abstract val modDescription: Property<String>
 
     @get:Input
-    /** Search tags. */
-    abstract val tags: ListProperty<String>
+    /** Project categories. */
+    abstract val categories: Property<TerracottaProjectCategories>
 
     @get:Input
     /** SPDX license identifier. */
@@ -152,7 +153,7 @@ abstract class TerracottaPlanTask : DefaultTask() {
             summary = summary.get(),
             description = modDescription.get(),
             versions = listOf(version),
-            tags = tags.get(),
+            categories = categories.get(),
             license = license.get(),
             licenseUrl = licenseUrl.orNull,
             icon = icon.orNull?.asFile?.absolutePath,